Extract all of github
